+++ index.js
@@ -2,13 +2,16 @@
 
 var browserify = require('browserify');
 var Promise = require('promise');
+var path = require('path');
+var Readable = require('stream').Readable;
+var extend = require('extend-shallow');
 
 exports.name = 'browserify';
 exports.outputFormat = 'js';
 
 exports.renderFileAsync = function (filename, options, locals) {
   return new Promise(function (resolve, reject) {
-    var b = browserify(options);
+    var b = browserify(options || {});
     b.add(filename, locals);
     b.bundle(function (err, buf) {
       if (err) {
@@ -23,12 +26,25 @@ exports.renderFileAsync = function (filename, options, locals) {
 
 exports.renderAsync = function (str, options, locals) {
   return new Promise(function (resolve, reject) {
-    var browserifyString = require('browserify-string');
-    browserifyString(str, options).bundle(function (err, src) {
+    // Inject the basedir from the filename, if available.
+    options = extend({}, options, locals);
+    if (options.hasOwnProperty('filename') && !options.hasOwnProperty('basedir')) {
+      options.basedir = path.dirname(options.filename);
+    }
+
+    // Create a stream from the input string.
+    var stream = new Readable();
+    stream.push(str);
+    stream.push(null); // Indicate the end of the stream.
+
+    // Create the Browserify object.
+    var b = browserify(stream, options);
+    b.bundle(function (err, buf) {
       if (err) {
         reject(err);
-      } else {
-        resolve(src.toString());
+      }
+      else {
+        resolve(buf.toString());
       }
     })
   })